+++ src/type/global.d.ts
@@ -1,9 +1,11 @@
 import type { Mermaid } from 'mermaid'
 import type { Flexsearch } from 'flexsearch'
+import JSZip from 'jszip'
 
 declare global {
   interface Window {
     FlexSearch: Flexsearch
+    JSZip: typeof JSZip
     mermaid: Mermaid
     goToHref: (event: MouseEvent, href: string) => void
   }