+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## unreleased
 
+- change: externalize jszip
 - fix: stat addNumeric() `item[attribut.variable]` must be true not just type of string
 - fix: evolution col variable with `<br>`
+- fix: evolution variable two rows and null value
 
 ## 0.13.0 (2024-10-02)